### 2024-03-30 #636 @mikemahoney218
rsi: Efficiently Retrieve and Process Satellite Imagery
Submitting Author Name: Mike Mahoney
Submitting Author Github Handle: @mikemahoney218
Repository:  https://github.com/Permian-Global-Research/rsi/
Version submitted:
Submission type: Standard

Version accepted: TBD

---

```
Package: rsi
Title: Efficiently Retrieve and Process Satellite Imagery
Version: 0.2.0.9000
    person("Michael", "Mahoney", , "mike.mahoney.218@gmail.com", role = c("aut", "cre"),
           comment = c(ORCID = "0000-0003-2402-304X")),
    person("Permian Global", role = c("cph", "fnd"))
  )
Description: Downloads spatial data from spatiotemporal asset catalogs
    ('STAC'), computes standard spectral indices from the Awesome Spectral
    Indices project (Montero et al. (2023) <doi:10.1038/s41597-023-02096-0>)
    against raster data, and glues the outputs together into predictor bricks.
    Methods focus on interoperability with the broader spatial ecosystem;
    function arguments and outputs use classes from 'sf' and 'terra', and data
    downloading functions support complex 'CQL2' queries using 'rstac'.
License: Apache License (>= 2)
    future.apply,
    glue,
    httr,
    lifecycle,
    proceduralnames,
    rstac,
    sf,
    terra,
    tibble
    curl,
    progressr,
    withr
Config/testthat/parallel: true
```

This package supports (spatial) data retrieval from APIs implementing the OGC STAC API standard, processing the downloaded data (including automated masking, compositing and rescaling) computing spectral indices from those data, and wrangling the outputs into formats useful for modeling and visualization.

Anyone with a need to download and process spatial data, particularly remote sensing data, particularly satellite-based earth observation rasters. We've used rsi to automate the entire data preparation process of forest carbon and structure models (not yet published), but the package is broadly useful to anyone working in Earth surface modeling.

Yes:

- [rstac](https://github.com/brazil-data-cube/rstac) is a fantastic package for querying and downloading data from STAC APIs. rstac does not implement the other elements of rsi (compositing, masking, rescaling, computing spectral indices, wrangling the outputs), and rsi provides a "higher level" method for downloading data from STAC APIs (powered by lower-level rstac functions); rsi also uses a faster download method.
- [gdalcubes](https://gdalcubes.github.io/) and [sits](https://e-sensing.github.io/sitsbook/) both provide higher-level approaches for accessing data from STAC APIs, organized around data cube models. There is more substantial overlap between these packages and rsi. The biggest difference is that rsi, very intentionally, does not provide a new data model. To quote [the vignette](https://permian-global-research.github.io/rsi/articles/Downloading-data-from-STAC-APIs-using-rsi.html#a-quick-note-on-rsi-and-other-packages-for-downloading-stac-assets):

> A core difference between rsi and these packages is that rsi does not have a data model: rsi is focused entirely on finding the bits of data you want from remote endpoints, and getting those bits on your local machine for you to process with your normal spatial data tooling. There are no new classes in rsi (other than the band mapping objects), and the outputs of functions are local rasters. This is an approach that fits better in my head than the more abstract delayed computations in some other packages; at the same time, it’s possible that this approach can be less efficient, downloading more data at finer resolutions than is actually needed for a given task.

There are a few minor things that I think rsi does better than other approaches (we sign items right before each one is downloaded, for instance, whereas some other packages sign items before starting to download the entire set, meaning the signature can expire causing large downloads to fail), but this difference I think is mostly a matter of taste. I'm very familiar with local GDAL, terra, and sf, and so rsi tries to get users back to working with local GDAL, terra, and sf as fast as possible.

NA

NA

~I have never successfully gotten the CI item to be a check, and I have no idea why. I'm using the standard `usethis` functions to structure my CI, for what it's worth!~ Apparently this is a local-only issue.

I addressed failing covr in https://github.com/ropensci/software-review/issues/636#issuecomment-2028086514
