### 2024-03-30 #636 @mikemahoney218
Another quick note -- I will not be able to transfer this repository to rOpenSci if accepted. I had asked on Slack and was told by Yani that this was acceptable, [though it's only mentioned in the book here](https://devguide.ropensci.org/softwarereview_editor.html#after-review). I want to flag this at the start, in case it turns out to be an issue!
