### 2024-03-30 #636 @mikemahoney218
Guessing covr fails due to not setting my custom `is_covr` environment variable:
https://github.com/Permian-Global-Research/rsi/commit/c62e6e9aa08c06a54094edc6c3df994260d8aa17

This environment variable is used to skip a test on my covr CI. The tl;dr is that [rsi executes some code in a minimal environment to protect against malicious code downloaded from the internet](https://www.mm218.dev/posts/2023-10-27-minimal-environments/), which prevents covr from injecting its tracking inside of that minimal environment. I still want the file to be tested (and the other pieces of the file to be counted in coverage), though, so I wrapped the local environment section in `nocov` and added this environment variable.

You can see my code coverage report at https://app.codecov.io/gh/Permian-Global-Research/rsi and my CI workflow for this at https://github.com/Permian-Global-Research/rsi/blob/main/.github/workflows/test-coverage.yaml
