+++ dev/pkgdown.yml
@@ -5,7 +5,7 @@ articles:
   Downloading-data-from-STAC-APIs-using-rsi: Downloading-data-from-STAC-APIs-using-rsi.html
   How-can-I-: How-can-I-.html
   rsi: rsi.html
-last_built: 2024-03-29T15:13Z
+last_built: 2024-03-30T14:36Z
 urls:
   reference: https://permian-global-research.github.io/rsi/reference
   article: https://permian-global-research.github.io/rsi/articles
