+++ CITATION.cff
@@ -8,7 +8,8 @@ message: 'To cite package "rsi" in publications use:'
 type: software
 license: Apache-2.0
 title: 'rsi: Efficiently Retrieve and Process Satellite Imagery'
-version: 0.1.2.9000
+version: 0.2.0.9000
+doi: 10.5281/zenodo.10926857
 abstract: Downloads spatial data from spatiotemporal asset catalogs ('STAC'), computes
   standard spectral indices from the Awesome Spectral Indices project (Montero et
   al. (2023) <doi:10.1038/s41597-023-02096-0>) against raster data, and glues the
@@ -21,6 +22,17 @@ authors:
   given-names: Michael
   email: mike.mahoney.218@gmail.com
   orcid: https://orcid.org/0000-0003-2402-304X
+preferred-citation:
+  type: generic
+  title: 'rsi: Efficiently Retrieve and Process Satellite Imagery'
+  authors:
+  - family-names: Mahoney
+    given-names: Michael
+    email: mike.mahoney.218@gmail.com
+    orcid: https://orcid.org/0000-0003-2402-304X
+  url: https://permian-global-research.github.io/rsi/
+  doi: 10.5281/zenodo.10926857
+  version: 0.2.0.9000
 repository: https://CRAN.R-project.org/package=rsi
 repository-code: https://github.com/Permian-Global-Research/rsi
 url: https://permian-global-research.github.io/rsi/
