### 2024-04-05 17dc778
Clarify calculating indices based on available data
Fixes #60
