### 2024-04-05 e71186f
Clarify calculating indices based on available data (#61)
Fixes #60
