+++ dev/pkgdown.yml
@@ -5,7 +5,7 @@ articles:
   Downloading-data-from-STAC-APIs-using-rsi: Downloading-data-from-STAC-APIs-using-rsi.html
   How-can-I-: How-can-I-.html
   rsi: rsi.html
-last_built: 2024-04-04T14:01Z
+last_built: 2024-04-05T20:22Z
 urls:
   reference: https://permian-global-research.github.io/rsi/reference
   article: https://permian-global-research.github.io/rsi/articles
