### 2024-04-11 #636 @mikemahoney218
Just want to flag that I'm still expecting (your version of) codecov to fail, due to https://github.com/ropensci/software-review/issues/636#issuecomment-2028086514

The core issue is that `calculate_indices()` is basically intended to run code from a random site on the internet -- a trusted site, but still a security risk. As such, that downloaded code is run inside a minimal environment that prevents injecting any unexpected code or functions. Unfortunately, covr works by injecting unexpected functions into your source code, and then counting how many times its functions get run -- then `calculate_indices()` doesn't allow those to execute, causing the function to fail.

As a result, I toggle the tests that hit this code path using a custom `is_covr` variable, which isn't set by your covr check and so your covr check fails.

I don't want to disable the whole .R file from covr, because covr can instrument the rest of the file, and I don't want to drop these tests (or make them off by default) because I'd like `R CMD check` to check this function. I've got a live coverage report running via GHA and hopefully viewable at:
https://app.codecov.io/gh/Permian-Global-Research/rsi?branch=main
