### 2024-04-11 #636 @jhollist
@mikemahoney218 Thanks for the update.  And as you expected the bot checks fail.  I am not too worried about that given you have good coverage and can demonstrate it with the codecov reports.  We will want to make sure that this coverage is reported out in the repositories README.  You may already be doing that, I just haven't checked yet!

Stay tuned, I am working on this today and tomorrow and expect to move on to finding reviewers shortly after that!
