### 2024-04-11 #636 @jhollist
---

I think we are ready to pass on to reviewers!  Nice Job!

Only one very small request:

- Add a link to the CONTRIBUTING.md file on the README.

I like to see a more upfront CONTRIBUTING.  I always get lost trying to find them when embedded inside `.github`.  A simple link to that file should suffice!

Also, I have no concerns about the tests failing on the bot.  You have implemented them well, the coverage is good, and the badge makes it easy to find.
