### 2024-04-11 #636 @mikemahoney218
Added the link to CONTRIBUTING!
