### 2024-04-12 feee480
Check that it's _our_ error that fires
