+++ NEWS.md
@@ -1,5 +1,8 @@
 # rsi (development version)
 
+* `calculate_indices()` gains a new argument, `wopt`, which is passed directly
+  to `terra::predict()`.
+
 * Added a new section to the "How can I?" article on the pkgdown site, with
   pointers on how to "Calculate all possible indices using a certain data set".
   Thanks to @alkimj in #60 (#61).
