### 2024-06-05 512505a
Add cores
