### 2024-06-05 145745b
Cores (#66)
* Add wopt

* Add cores

* Add NEWS bullet
