### 2024-06-07 2b898ec
Add extra_objects
