### 2024-06-11 23c8f6d
Add check_crs argument
