### 2024-06-24 dd00d29
Add c.rsi_band_mapping method
