+++ tests/testthat/_snaps/misc.md
@@ -43,3 +43,14 @@
       red 
       "R" 
 
+# c works
+
+    Code
+      c(rsi::sentinel2_band_mapping$planetary_computer_v1, scl = "scl")
+    Output
+      An rsi band mapping object with attributes:
+      mask_band mask_function stac_source collection_name query_function class scl_name download_function sign_function names
+      
+        B01   B02   B03   B04   B05   B06   B07   B08   B8A   B09   B11   B12   scl 
+        "A"   "B"   "G"   "R" "RE1" "RE2" "RE3"   "N"  "N2"  "WV"  "S1"  "S2" "scl" 
+
