### 2024-06-24 145596f
Add c.rsi_band_mapping method (#72)
