### 2024-06-27 8ca7dd1
Remove bad install instructions
