+++ DESCRIPTION
@@ -1,6 +1,6 @@
 Package: rsi
 Title: Efficiently Retrieve and Process Satellite Imagery
-Version: 0.2.0.9000
+Version: 0.2.1
 Authors@R: c(
     person("Michael", "Mahoney", , "mike.mahoney.218@gmail.com", role = c("aut", "cre"),
            comment = c(ORCID = "0000-0003-2402-304X")),
