+++ NEWS.md
@@ -1,3 +1,5 @@
+# rsi (development version)
+
 # rsi 0.2.1
 
 * `calculate_indices()` gains several new arguments:
