### 2024-07-08 be0da7b
Survive past failed download when possible
