+++ R/get_stac_data.R
@@ -364,6 +364,9 @@ get_stac_data <- function(aoi,
     gdal_config_options = gdal_config_options,
     ...
   )
+  if (!is.null(na.action(download_results))) {
+    items$features[na.action(download_results)] <- NULL
+  }
   # mask
   if (!is.null(mask_band)) {
     download_results <- rsi_apply_masks(
