+++ R/get_stac_data.R
@@ -364,8 +364,8 @@ get_stac_data <- function(aoi,
     gdal_config_options = gdal_config_options,
     ...
   )
-  if (!is.null(na.action(download_results))) {
-    items$features[na.action(download_results)] <- NULL
+  if (!is.null(stats::na.action(download_results))) {
+    items$features[stats::na.action(download_results)] <- NULL
   }
   # mask
   if (!is.null(mask_band)) {
