### 2024-07-08 bcf8a26
Always assign na_attr
