### 2024-07-08 3718eeb
Survive past failed download when possible (#75)
* Survive past failed download when possible

* Namespace na.action

* Always assign na_attr
