+++ dev/pkgdown.yml
@@ -1,12 +1,11 @@
 pandoc: 3.1.11
-pkgdown: 2.0.9
+pkgdown: 2.1.0
 pkgdown_sha: ~
 articles:
-  Downloading-data-from-STAC-APIs-using-rsi: Downloading-data-from-STAC-APIs-using-rsi.html
-  How-can-I-: How-can-I-.html
+  articles/Downloading-data-from-STAC-APIs-using-rsi: Downloading-data-from-STAC-APIs-using-rsi.html
+  articles/How-can-I-: How-can-I-.html
   rsi: rsi.html
-last_built: 2024-06-28T14:59Z
+last_built: 2024-07-08T19:34Z
 urls:
   reference: https://permian-global-research.github.io/rsi/reference
   article: https://permian-global-research.github.io/rsi/articles
-
