### 2024-07-09 #636 @OldLipe
# {rsi} Package Review

-   [x] As the reviewer I confirm that there are no [conflicts of interest](https://devguide.ropensci.org/policies.html#coi) for me to review this work (if you are unsure whether you are in conflict, please speak to your editor *before* starting your review).

Thank you for the opportunity to review the `rsi` package. This package provides valuable functions to downloading satellite imagery using STAC catalogs.

I conducted the review from the perspective of both an end-user and a developer. I believe that both audiences should be addressed by the package.

To reproduce, I used the version `1.4.2-3` of Docker image of the [SITS](https://hub.docker.com/r/brazildatacube/sits-rstudio) package, with R version 4.3.0 and GDAL 3.0.4.

The documentation is comprehensive and covers all functions of the `rsi` package, including examples of how to use them.

1- I have only one suggestion regarding the parameters `gdalwarp_options` and `gdal_config_options`, which have very large default values. This somewhat clutters the code documentation. Could these GDAL values be stored in a configuration file, such as `.json` or `.yaml`, or even as environment variables? Using external configuration files allows for better organization of the code, and thus, you can remove these two parameters from your functions.

### Examples

All tested examples worked correctly. However, I have some questions that I listed below:

1- In the example of the `alos_palsar_mask_function()`, two warnings appeared. Listed below:

```R
aoi <- sf::st_point(c(-74.912131, 44.080410))
aoi <- sf::st_set_crs(sf::st_sfc(aoi), 4326)
aoi <- sf::st_buffer(sf::st_transform(aoi, 5070), 100)

palsar_image <- get_alos_palsar_imagery(
  aoi,
  start_date = "2021-01-01",
  end_date = "2021-12-31",
  mask_function = alos_palsar_mask_function
)

> Warning messages:
> 1: In CPL_gdalwarp(source, destination, options, oo, doo, config_options,  :
>   GDAL Message 1: Setting nodata to nan on band 2, but band 1 has nodata at nan. The > TIFFTAG_GDAL_NODATA only support one value per dataset. This value of nan will be used for all bands > on re-opening
> 2: In CPL_gdalwarp(source, destination, options, oo, doo, config_options,  :
>   GDAL Message 1: Setting nodata to nan on band 3, but band 1 has nodata at nan. The > TIFFTAG_GDAL_NODATA only support one value per dataset. This value of nan will be used for all bands > on re-opening
```

Is this behavior expected? Additionally, in this example, wouldn't it be more appropriate to save the image in a temporary directory? From a user's perspective, these warnings might scare new users.

2- The example in the `calculate_indices()` function presents an error. I understand that you are demonstrating an example that does not work. Users sometimes enter the documentation, copy the code, execute it, and then test it. Instead of the error, perhaps providing more examples of how to use the function would be helpful.

```R
example_indices <- filter_platforms(platforms = "Sentinel-1 (Dual Polarisation VV-VH)")[1, ]
example_indices$formula <- 'system("echo something bad")'
try(
  calculate_indices(
    system.file("rasters/example_sentinel1.tif", package = "rsi"),
    example_indices,
    tempfile(fileext = ".tif")
  )
)

> Error in system("echo something bad") : could not find function "system"
```

3- The functions that access catalogs: `get_stac_data()`, `get_sentinel1_imagery()`, `get_sentinel2_imagery()`, `get_landsat_imagery()`, `get_naip_imagery()`, `get_alos_palsar_imagery(),` `get_dem()` share the same documentation. This is indeed an excellent solution and saves time for both the user and the developer. To highlight the features of the package, one recommendation is to include new examples for other types of data, such as radar, since the names of the assets vary for each collection.

### Code overview

The code is well organized and well developed. Additionally, it has excellent coverage. I have some suggestions regarding the default values of certain parameters in your functions.

1- By default, the functions that download satellite images use random names for the images. I believe that using random names for satellite images could pose challenges for end users.

Here are a few points to consider:

- One method to index metadata for satellite images in relational databases is by utilizing metadata available within the filenames. Filenames are used to identify details such as satellite, sensor, band, grid (tile), and the date of the scene.
- For end users, this could lead to confusion, especially in identifying which image has already been downloaded.
- Standardizing the names of images allows for implementing recovery in downloads, in case they are interrupted due to external reasons.

I understand that some providers do not standardize the names of assets. However, your implementation allows you to create a standard filename for each collection, given that there are separate functions for each. I believe a good default approach is to maintain the original asset names from the providers.

2- I find the idea of simplifying the use of masks and allowing users to insert a new function quite interesting. However, I have some questions about the default values that were used:

- In the `sentinel2_mask_function()` function, I would like to understand why the value `2` (`shadows`) is kept as a non-masked value?

- Regarding Landsat, I believe there could be improvements in filtering the values to be retained. Additionally, the implementation could be easier if working with bit values rather than integers. With integers, we are limited to only a few combinations.

Consider the values for the QA_PIXEL band:

```
0 : "Missing data"
1 : "Dilated Cloud"
2 : "High confidence cirrus"
3 : "High confidence cloud"
4 : "High confidence cloud shadow"
5 : "High confidence snow cover"
6 : "Clear"
7 : "Water"
8 : "Low/High confidence of cloud"
9 : "Medium/High confidence of cloud"
10: "Low/High confidence of cloud shadow"
11: "Medium/High confidence of cloud shadow"
12: "Low/High confidence of snow"
13: "Medium/High confidence of snow"
14: "Low/High confidence of cirrus"
15: "Medium/High confidence of cirrus"
```

Using bits allows for a wide range of combinations to filter values. One package that implements this approach is SITS. SITS produces official land use and land cover data for the Brazilian states. See an implementation example using bit wise operator:

```R
interp_values <- c(0, 1, 2, 3, 4, 5, 9, 11, 13, 15)
bitwAnd(values, sum(2^interp_values))
```

Here, `values` represent the image values. This method makes it easier to filter the numerous combinations that may exist when using this band.

### Filter functions

I find these filtering functions interesting, but I believe they may not be scalable. For instance, you provide two filtering functions: `filter_platforms()` and `filter_bands()`. What if a user wants to filter by cloud percentage within items? Or apply more complex filters to the properties of each item?

I understand that you provide an `item_filter_function` parameter, but in all these filtering functions, you filter values using `vapply`. The `rstac` package includes a function `items_filter()` to facilitate item filtering. I believe using this function would simplify the approach to applying more complex filters based on items.

### Minor comment

I imagine that in the future your `calculate_indices()` function will support expressions through the dots parameter (`...`). I'd like to provide an example of implementation from the SITS package. The `sits_apply()` function allows users to generate indices using a provided high-level expression, like this:

```R
sits_apply(NDVI_norm = (NDVI - min(NDVI)) / (max(NDVI) - min(NDVI)))
```

This approach enables users to utilize native R functions such as `max`, `min`, and `mean`.
