### 2024-07-23 5adacb2
docs: use functions to set config values
