+++ NAMESPACE
@@ -19,6 +19,8 @@ export(get_stac_data)
 export(landsat_mask_function)
 export(landsat_platform_filter)
 export(rsi_download_rasters)
+export(rsi_gdal_config_options)
+export(rsi_gdalwarp_options)
 export(rsi_query_api)
 export(sentinel2_mask_function)
 export(sign_planetary_computer)
