### 2024-08-12 #636 @mdsumner
- [x] As the reviewer I confirm that there are no [conflicts of interest](https://devguide.ropensci.org/policies.html#coi).

- [ x **Automated tests:** Unit tests cover essential functions of the package and a reasonable range of inputs and conditions. All tests pass on the local machine.

---

I really like this package, I see a lot of familiar experiences and the result is a good and consistent set of functions for navigating this space. I haven't personally done anything with spectral indices, usually I expore imagery and how "it looks". I appreciate having all this ease-of-use tooling at hand, and I will be pointing colleagues directly to this package, and I hope to stay involved in at least small ways.

Please make mention somewhere that "where the compute runs" (i.e. what "local" means) is important here. There's an issue with performance *in different regions*  - and some pointers to how one might run in a region closer to where the usual data sources are (us-west2 for example. I'm not suggesting that's an easy topic to cover, just would like to see it mentioned.  It takes about 2x as long to run examples here in Tasmania, vs computers in us-west (I know that comparison is a lot more complex than this).  It might be an idea to point to ways of running computers on public systems that run closer to the data, or at least a guide to that.

I'm a bit disappointed in our community in that packages in R have tended towards doing "everything", and (not a criticism of this package) here we are doing a  lot of things: getting file sources from STAC (a json web query),  configuring for authentication,  downloading specific files (a curl task), mosaicing and standardizing imagery (a GDAL task). I like that the top level functions are decomposed in this package itself, and the documentation is clear as to what underlying functions are used, and I can understand why parts of "foundational" packages are used with a mix of underlying generic tools for this complex stack of work.  I'm massively impressed with how much this package actually does, and exposes in a general way down to the level of templated types of functionality and details like the GDALWarp options.

rsi is a little bit like a datacube, but with exposure of the underlying components at key steps. I really like how the asset is a lightly classed object with other functions that it needs stored as attributes, and that that is how the remaining arguments to the getter function is structured, so it seamlessly inherits each level but also remains flexible to user-changes in that one call.

####  Questions I had that I could put more effort+examples into:

I'd like see some validations of the raster values obtained in some examples, if there was an external example that's documented elsewhere (in a python notebook, or another R package) it would be neat to have a clear comparison of the scale/s shown be these raster files obtained here against an independent example. (I had intended to do this myself ...)

I was looking for an example where I can make a true colour image with RGB, I don't understand the scaling that occurs in the examples (we have 0-1 scale numbers, which don't work with terra::plotRGB or its stretch argument off the shelf, and leave me unclear about the scaled ranges and whether I am plotting an RGB image correctly. The first example in the readme plots an RGB image as separate bands, and I'm unclear what to do to plot that as an "image".   Again, I had meant to provide an actual example here.   I believe the advice should be:

```
plotRGB(stretch(x))

## note that  this provides different defaults to stretch() itself
plotRGB(x, stretch = "lin")
```

but also, it's open to interpretation and expert use afaics.

#### Specific notes

Two of these three files have no data in the aoi region.

```R
qfiles <- get_landsat_imagery(aoi,
 start_date = "2022-06-01", end_date = "2022-06-30",
composite_function = NULL)
```

I tried this naive thing, to run rsi_query_api, and I think at the least it should error fast when not getting a bbox in longlat.

```
rsi_query_api(  aoi,
+                 start_date = "2022-06-01",
+                 end_date = "2022-06-30",)
Error in rsi_query_api(aoi, start_date = "2022-06-01", end_date = "2022-06-30",  :
  argument "stac_source" is missing, with no default
> rsi_query_api(  aoi, stac_source = "https://planetarycomputer.microsoft.com/api/stac/v1/",
+                 start_date = "2022-06-01",
+                 end_date = "2022-06-30",)
Error in rsi_query_api(aoi, stac_source = "https://planetarycomputer.microsoft.com/api/stac/v1/",  :
  argument "collection" is missing, with no default
> rsi_query_api(  aoi, stac_source = "https://planetarycomputer.microsoft.com/api/stac/v1/",
+                 start_date = "2022-06-01",
+                 end_date = "2022-06-30", collection = "sentinel2-c1-l2a")
Error in Ops.sfg(bbox[[2]], bbox[[4]]) :
  operation > not supported for sfg objects
> rsi_query_api(  sf::st_bbox(aoi), stac_source = "https://planetarycomputer.microsoft.com/api/stac/v1/",
+                 start_date = "2022-06-01",
+                 end_date = "2022-06-30", collection = "sentinel2-c1-l2a")
Error in rsi_query_api(sf::st_bbox(aoi), stac_source = "https://planetarycomputer.microsoft.com/api/stac/v1/",  :
  argument "limit" is missing, with no default
> rsi_query_api(  sf::st_bbox(aoi), stac_source = "https://planetarycomputer.microsoft.com/api/stac/v1/",
+                 start_date = "2022-06-01",
+                 end_date = "2022-06-30", collection = "sentinel2-c1-l2a", limit  = 100)

```

it could detect that the bbox input is not in longlat and 1) error or 2) just transform it.

> It's often useful to buffer your aoi object slightly, on the order of 1-2 cell widths, in order to ensure that data is downloaded for your entire AOI even after accounting for any reprojection needed to compare your AOI to the data on the STAC server.

Here I would rather reproject the extent, this is not so hard to do and exists in a few places
raster::projectExtent, reproj::reproj_extent, and terra::project but possibly the best option is to use GDAL warp (via sf as elsewhere here) to reproject an empty raster.  (Happy to follow up to illustrate).

> This function can either download all data that intersects with your spatiotemporal AOI as multiple files (if composite_function = NULL), or can be used to rescale band values, apply a mask function, and create a composite from the resulting files in a single function call

On this point, GDAL warp can itself do these things, itself a monolith of abstractions itself but we can possibly avoid downloading entire scenes. I mention this not as a strong suggestion, mainly an invite to discuss further (I'm looking at the rise of {gdalraster} here).

>  It can be a good idea to tile your aoi using sf::st_make_grid()

I think this really needs an example, because there's room for guidance on creating a nice tile mosaic definition (with terra::align for example, or actually with st_tile or getTileExtents). It's very important point, and say what if we wanted a (CONUS) state-level imagery. This can be done a little more abstractly using terra::makeTiles and stars::st_tile (which give the index and extents helpfully but separately).  (Maybe an assign-to-me task).

> If you set the rsi_pc_key environment variable to your key (either primary or secondary; there is no difference), rsi will automatically use this key to sign all requests against Planetary Computer.

I only note this as a discussion-bounce, for possible followup: GDAL can do this too, and has file system abstrations that can copy, info, or warp-in-part to a target grid.

In the vignette I like when there's "one change", that's an excellent situation when you can tweak major changes with only one tiny plumbing modification.

This example (in the README) should save the file name as an output. It's otherwise a pipeline that I don't get an object for in the end, and it can take some time (in Australia).

```R
projected_ashe <- sf::st_transform(ashe, 6542)
> get_landsat_imagery(
+     aoi = projected_ashe,
+     start_date = "2021-06-01",
+     end_date = "2021-06-30",
+     output_filename = tempfile(fileext = ".tif")
+ ) |>
+     terra::rast() |>
+     terra::plot()
```

I have some minor discomforts about when exactly warping or masking is done. I just want to talk about the details of this and might follow up, I'm a bit lost in the depths of the compositing function and use of sprc and mosaic, though.

A standalone question I had: can we use the aoi to drive the STAC query, but then download the files as-is without cropping/warping or compositing?  I guess that would mean providing a link to the temp-file space being used.

Fin. Thanks so much for such a great package, and much apology for being so late here (especially appreciative of the patience of @mikemahoney218 and @jhollist).
