### 2024-08-13 b1a15ba
Update alos_palsar_mask_function() documentation
