### 2024-08-13 e2ec137
Provide more examples in calculate_indices()
