### 2024-08-13 7b82f3d
Add documentation to get_*_imagery()
