+++ R/get_stac_data.R
@@ -181,6 +181,35 @@
 #'   start_date = "2022-06-01",
 #'   end_date = "2022-08-30"
 #' )
+#' 
+#' # The `get_*_imagery()` functions will download 
+#' # all available "data" assets by default
+#' # (usually including measured values, and excluding derived bands)
+#' sentinel1_data <- get_sentinel1_imagery(
+#'   aoi,
+#'   start_date = "2022-06-01",
+#'   end_date = "2022-07-01"
+#' )
+#' names(terra::rast(sentinel1_data))
+#' 
+#' # You can see what bands will be downloaded by a function
+#' # by inspecting the corresponding `band_mapping` object:
+#' sentinel2_band_mapping$planetary_computer_v1
+#' 
+#' # And you can add additional assets using `c()`:
+#' c(
+#'   sentinel2_band_mapping$planetary_computer_v1,
+#'   "scl"
+#' )
+#' 
+#' # Or subset the assets downloaded using `[` or `[[`:
+#' sentinel2_imagery <- get_sentinel2_imagery(
+#'   aoi,
+#'   start_date = "2022-06-01",
+#'   end_date = "2022-07-01",
+#'   asset_names = sentinel2_band_mapping$planetary_computer_v1["B01"]
+#' )
+#' names(terra::rast(sentinel2_imagery))
 #'
 #' @export
 get_stac_data <- function(aoi,
