### 2024-08-13 #636 @mikemahoney218
Thank you so much, @mdsumner ! I'm excited to dive into your review.

As I said to Jeff over email, I also haven't been a paragon of timeliness here -- I got started replying to Felipe [three weeks ago, and then immediately lost my focus](https://github.com/Permian-Global-Research/rsi/commits/old_lipe/). I'm hoping to carve out time to respond to both reviews starting on Wednesday of this week!
