### 2024-08-15 f499c95
Add masked_bits() argument
