+++ R/mask_functions.R
@@ -21,7 +21,6 @@ sentinel2_mask_function <- function(raster) {
   terra::`%in%`(
     raster,
     c(
-      2, # DARK_AREA_PIXELS
       4, # VEGETATION
       5, # NOT_VEGETATED
       6, # WATER
@@ -33,6 +32,7 @@ sentinel2_mask_function <- function(raster) {
   # c(
   # 0, # NO_DATA
   # 1, # SATURATED_OR_DEFECTIVE
+  # 2, # DARK_AREA_PIXELS
   # 3, # CLOUD_SHADOWS
   # 8, # CLOUD_MEDIUM_PROBABILITY
   # 9, # CLOUD_HIGH_PROBABILITY
@@ -45,6 +45,8 @@ sentinel2_mask_function <- function(raster) {
 #' @param raster The QA band of a Landsat image
 #' @param include Include pixels that represent land, water, or both? Passing
 #' `c("land", "water")` is identical to passing `"both"`.
+#' @inheritParams rlang::args_dots_empty
+#' @param masked_bits 
 #'
 #' @returns A boolean raster to be used to mask a Landsat image
 #'
@@ -57,24 +59,73 @@ sentinel2_mask_function <- function(raster) {
 #'   aoi,
 #'   start_date = "2022-06-01",
 #'   end_date = "2022-08-30",
-#'   mask_function = landsat_mask_function
+#'   mask_function = landsat_mask_function,
+#'   output_file = tempfile(fileext = ".tif")
+#' )
+#' 
+#' # Or, optionally pass the qa_pixel bits to mask out directly
+#' landsat_image <- get_landsat_imagery(
+#'   aoi,
+#'   start_date = "2022-06-01",
+#'   end_date = "2022-08-30",
+#'   mask_function = \(x) landsat_mask_function(
+#'     x, 
+#'     masked_bits = list(c(0:5, 7, 9, 11, 13, 15))
+#'   ),
+#'   output_file = tempfile(fileext = ".tif")
+#' ) 
+#' 
+#' # You can use this to specify multiple acceptable values
+#' # from the qa_pixel bitmask; names are optional
+#' landsat_image <- get_landsat_imagery(
+#'   aoi,
+#'   start_date = "2022-06-01",
+#'   end_date = "2022-08-30",
+#'   mask_function = \(x) landsat_mask_function(
+#'     x, 
+#'     masked_bits = list(
+#'       clear_land = c(0:5, 7, 9, 11, 13, 15),
+#'       clear_water = c(0:5, 9, 11, 13, 15)
+#'     )
+#'   ),
+#'   output_file = tempfile(fileext = ".tif")
 #' )
 #'
 #' @export
-landsat_mask_function <- function(raster, include = c("land", "water", "both")) {
-  if (missing(include)) include <- include[[1]]
-  include <- rlang::arg_match(include, multiple = TRUE)
-  classes <- numeric()
+landsat_mask_function <- function(raster, 
+                                  include = c("land", "water", "both"),
+                                  ..., 
+                                  masked_bits) {
+  rlang::check_dots_empty()
+  if (missing(masked_bits)) {
+    if (missing(include)) include <- include[[1]]
+    include <- rlang::arg_match(include, multiple = TRUE)
 
-  # "Clear with lows set"
-  # https://d9-wret.s3.us-west-2.amazonaws.com/assets/palladium/production/s3fs-public/media/files/LSDS-1619_Landsat-8-9-C2-L2-ScienceProductGuide-v4.pdf
-  if (any(c("land", "both") %in% include)) classes <- c(classes, 21824)
-  # from https://github.com/Permian-Global-Research/rsi/issues/37
[…]
