+++ man/landsat_mask_function.Rd
@@ -4,13 +4,22 @@
 \alias{landsat_mask_function}
 \title{Create a Landsat mask raster from the QA band}
 \usage{
-landsat_mask_function(raster, include = c("land", "water", "both"))
+landsat_mask_function(
+  raster,
+  include = c("land", "water", "both"),
+  ...,
+  masked_bits
+)
 }
 \arguments{
 \item{raster}{The QA band of a Landsat image}
 
 \item{include}{Include pixels that represent land, water, or both? Passing
 \code{c("land", "water")} is identical to passing \code{"both"}.}
+
+\item{...}{These dots are for future extensions and must be empty.}
+
+\item{masked_bits}{}
 }
 \value{
 A boolean raster to be used to mask a Landsat image
@@ -28,7 +37,36 @@ landsat_image <- get_landsat_imagery(
   aoi,
   start_date = "2022-06-01",
   end_date = "2022-08-30",
-  mask_function = landsat_mask_function
+  mask_function = landsat_mask_function,
+  output_file = tempfile(fileext = ".tif")
+)
+
+# Or, optionally pass the qa_pixel bits to mask out directly
+landsat_image <- get_landsat_imagery(
+  aoi,
+  start_date = "2022-06-01",
+  end_date = "2022-08-30",
+  mask_function = \(x) landsat_mask_function(
+    x, 
+    masked_bits = list(c(0:5, 7, 9, 11, 13, 15))
+  ),
+  output_file = tempfile(fileext = ".tif")
+) 
+
+# You can use this to specify multiple acceptable values
+# from the qa_pixel bitmask; names are optional
+landsat_image <- get_landsat_imagery(
+  aoi,
+  start_date = "2022-06-01",
+  end_date = "2022-08-30",
+  mask_function = \(x) landsat_mask_function(
+    x, 
+    masked_bits = list(
+      clear_land = c(0:5, 7, 9, 11, 13, 15)
+      clear_water = c(0:5, 9, 11, 13, 15)
+    )
+  ),
+  output_file = tempfile(fileext = ".tif")
 )
 \dontshow{\}) # examplesIf}
 }
