### 2024-08-15 bb556d2
Document masked_bits
