+++ man/landsat_mask_function.Rd
@@ -19,7 +19,12 @@ landsat_mask_function(
 
 \item{...}{These dots are for future extensions and must be empty.}
 
-\item{masked_bits}{}
+\item{masked_bits}{Optionally, a list of integer vectors representing the
+individual bits to mask out. Each vector is converted to an integer
+representation, and then pixels with matching \code{qa_pixel} values
+are preserved by the mask. Refer to the Landsat science product guide for
+further information on what bit values represent for your platform of
+interest.}
 }
 \value{
 A boolean raster to be used to mask a Landsat image
@@ -47,11 +52,11 @@ landsat_image <- get_landsat_imagery(
   start_date = "2022-06-01",
   end_date = "2022-08-30",
   mask_function = \(x) landsat_mask_function(
-    x, 
+    x,
     masked_bits = list(c(0:5, 7, 9, 11, 13, 15))
   ),
   output_file = tempfile(fileext = ".tif")
-) 
+)
 
 # You can use this to specify multiple acceptable values
 # from the qa_pixel bitmask; names are optional
@@ -60,9 +65,9 @@ landsat_image <- get_landsat_imagery(
   start_date = "2022-06-01",
   end_date = "2022-08-30",
   mask_function = \(x) landsat_mask_function(
-    x, 
+    x,
     masked_bits = list(
-      clear_land = c(0:5, 7, 9, 11, 13, 15)
+      clear_land = c(0:5, 7, 9, 11, 13, 15),
       clear_water = c(0:5, 9, 11, 13, 15)
     )
   ),
