### 2024-08-15 #636 @mikemahoney218
## Response to @OldLipe

Thank you so, so much for your review here @OldLipe ! I feel like your comments have really improved the package. I'll walk through specific changes below, but first I wanted to ask about one remaining comment:

> 1- By default, the functions that download satellite images use random names for the images. I believe that using random names for satellite images could pose challenges for end users.

The issue I have with generating non-random names is that I feel users are in a better position than I am to know how to organize their files within a project. Users downloading multiple files are likely iterating through either collections, time ranges, or spatial areas of interest, and probably have a pre-existing idea of what distinguishes each file (and therefore would make for a good name). This is why, for instance, [the documentation shows examples of providing your own `output_filename` values when you know how your downloads are being "chunked"](https://permian-global-research.github.io/rsi/articles/How-can-I-.html#get-one-composite-per-year-month-or-other-interval).

So the idea behind random filenames is that it's something that is good enough for fast proof-of-concept "does this function work" tests, but is clearly _not_ good enough for "real" usage. I'm hoping to force users to come up with their own file name conventions, rather than accepting the default options. This is also why I _do_ handle the auto-naming when `composite_function = NULL`, because I am very confident that I know what differentiates each file (the datetime) in that situation.

I'm curious what you think about this reasoning. If anything, I'd lean towards removing the default filenames altogether, but I do think they're useful for quick evaluation of the package.

> 1- [R]egarding the parameters `gdalwarp_options` and `gdal_config_options` [...] Could these GDAL values be stored in a configuration file?

I moved these into functions: https://github.com/Permian-Global-Research/rsi/pull/77/commits/5adacb2bbe5cc8ab9cae57e7a57db585c7ab67d6

> 1- In the example of the `alos_palsar_mask_function()`, two warnings appeared. Listed below:
> [...]
> Is this behavior expected? Additionally, in this example, wouldn't it be more appropriate to save the image in a temporary directory? From a user's perspective, these warnings might scare new users.

I couldn't reproduce the warning (though I've seen it before, I just don't understand what triggers it) but I updated the documentation with a gdalwarp option to silence the warning: https://github.com/Permian-Global-Research/rsi/pull/77/commits/b1a15bab095e27cd242b45feb55704864dc96539

I also saved to a temporary file -- thanks! Because these examples don't run on CRAN, that one is a really easy thing to accidentally miss :sweat_smile: There were a few more of these here: https://github.com/Permian-Global-Research/rsi/pull/77/commits/fa7dc91c3d5eb86ac087d9cc62cf756ebc32b1ce

> 2- The example in the `calculate_indices()` function presents an error. I understand that you are demonstrating an example that does not work. Users sometimes enter the documentation, copy the code, execute it, and then test it. Instead of the error, perhaps providing more examples of how to use the function would be helpful.

I left the error in place, because I want to explain _why_ this function intentionally doesn't let users use arbitrary functions. But I added more documentation around it: a comment right above the `try()` to emphasize that the error is expected, and then another block below it to show how to work around this design: https://github.com/Permian-Global-Research/rsi/pull/77/commits/e2ec1378b50c0c756f84f502ec2298a9bc63a5c0

> 3- The functions that access catalogs [...] share the same documentation. [...] [I]nclude new examples for other types of data, such as radar, since the names of the assets vary for each collection.

I added additional examples to this document, using multiple data-retrieving functions and showing how to work with band mapping objects: https://github.com/Permian-Global-Research/rsi/pull/77/commits/7b82f3d115da15ef29d4d3415b1c462170b68212

>     * In the `sentinel2_mask_function()` function, I would like to understand why the value `2` (`shadows`) is kept as a non-masked value?

This was a straight-up mistake; thank you. Fixed here: https://github.com/Permian-Global-Research/rsi/pull/77/commits/f499c9507f8a06a318c42a904aa3538ccf779baf

>     * Regarding Landsat, I believe there could be improvements in filtering the values to be retained. Additionally, the implementation could be easier if working with bit values rather than integers.

Thank you for the examples here; without them, I absolutely never would have figured out how bitmasks work. I added a `masked_bits` argument to the Landsat mask function, which can take vectors of bits to mask out. The `include` argument now works by setting these bits: https://github.com/Permian-Global-Research/rsi/pull/77/commits/f499c9507f8a06a318c42a904aa3538ccf779baf

> I find these filtering functions interesting, but I believe they may not be scalable. For instance, you provide two filtering functions: `filter_platforms()` and `filter_bands()`. What if a user wants to filter by cloud percentage within items? Or apply more complex filters to the properties of each item?

Discussed in https://github.com/ropensci/software-review/issues/636#issuecomment-2218735327

Thank you again -- this was a massively helpful review. Let me know if I missed anything (or made anything worse by mistake :smile:)
