### 2024-08-15 9e50b37
Changes for OldLipe review (#77)
* docs: use functions to set config values

* Update alos_palsar_mask_function() documentation

* Provide more examples in calculate_indices()

* Add documentation to get_*_imagery()

* Save to tempfiles

* Add masked_bits() argument

* Document masked_bits
