+++ R/get_stac_data.R
@@ -140,9 +140,11 @@
 #' argument of [sf::gdal_utils()]. The same set of options are used for all
 #' downloaded data and the final output images; this means that some common
 #' options (for instance, `PREDICTOR=3`) may cause errors if bands are of
-#' varying data types.
+#' varying data types. The default values are provided by 
+#' [rsi_gdalwarp_options()].
 #' @param gdal_config_options Options passed to `gdalwarp` through the
-#' `config_options` argument of [sf::gdal_utils()].
+#' `config_options` argument of [sf::gdal_utils()]. The default values are
+#' provided by [rsi_gdal_config_options()].
 #' @param platforms The names of Landsat satellites to download imagery from.
 #' These do not correspond to the `platforms` column in [spectral_indices()];
 #' the default argument of `c("landsat-9", "landsat-8")` corresponds to
@@ -170,15 +172,48 @@
 #'   mask_band = "qa_pixel",
 #'   mask_function = landsat_mask_function,
 #'   item_filter_function = landsat_platform_filter,
-#'   platforms = c("landsat-9", "landsat-8")
+#'   platforms = c("landsat-9", "landsat-8"),
+#'   output_filename = tempfile(fileext = ".tif")
 #' )
 #'
 #' # or, mostly equivalently (will download more bands):
 #' landsat_image <- get_landsat_imagery(
 #'   aoi,
 #'   start_date = "2022-06-01",
-#'   end_date = "2022-08-30"
+#'   end_date = "2022-08-30",
+#'   output_filename = tempfile(fileext = ".tif")
 #' )
+#' 
+#' # The `get_*_imagery()` functions will download 
+#' # all available "data" assets by default
+#' # (usually including measured values, and excluding derived bands)
+#' sentinel1_data <- get_sentinel1_imagery(
+#'   aoi,
+#'   start_date = "2022-06-01",
+#'   end_date = "2022-07-01",
+#'   output_filename = tempfile(fileext = ".tif")
+#' )
+#' names(terra::rast(sentinel1_data))
+#' 
+#' # You can see what bands will be downloaded by a function
+#' # by inspecting the corresponding `band_mapping` object:
+#' sentinel2_band_mapping$planetary_computer_v1
+#' 
+#' # And you can add additional assets using `c()`:
+#' c(
+#'   sentinel2_band_mapping$planetary_computer_v1,
+#'   "scl"
+#' )
+#' 
+#' # Or subset the assets downloaded using `[` or `[[`:
+#' sentinel2_imagery <- get_sentinel2_imagery(
+#'   aoi,
+#'   start_date = "2022-06-01",
+#'   end_date = "2022-07-01",
+#'   asset_names = sentinel2_band_mapping$planetary_computer_v1["B01"],
+#'   output_filename = tempfile(fileext = ".tif")
+#' )
+#' names(terra::rast(sentinel2_imagery))
 #'
 #' @export
 get_stac_data <- function(aoi,
@@ -200,26 +235,8 @@ get_stac_data <- function(aoi,
                           output_filename = paste0(proceduralnames::make_english_names(1), ".tif"),
                           composite_function = c("merge", "median", "mean", "sum", "min", "max"),
                           limit = 999,
-                          gdalwarp_options = c(
-                            "-r", "bilinear",
-                            "-multi",
-                            "-overwrite",
-                            "-co", "COMPRESS=DEFLATE",
-                            "-co", "PREDICTOR=2",
-                            "-co", "NUM_THREADS=ALL_CPUS"
-                          ),
-                          gdal_config_options = c(
-                            VSI_CACHE = "TRUE",
-                            GDAL_CACHEMAX = "30%",
[…]
