+++ man/alos_palsar_mask_function.Rd
@@ -28,7 +28,12 @@ palsar_image <- get_alos_palsar_imagery(
   aoi,
   start_date = "2021-01-01",
   end_date = "2021-12-31",
-  mask_function = alos_palsar_mask_function
+  mask_function = alos_palsar_mask_function,
+  output_file = tempfile(fileext = ".tif"),
+  gdalwarp_options = c(
+    rsi::rsi_gdalwarp_options(),
+    "-srcnodata", "nan"
+  )
 )
 \dontshow{\}) # examplesIf}
 }

[…]
