+++ dev/pkgdown.yml
@@ -5,7 +5,7 @@ articles:
   articles/Downloading-data-from-STAC-APIs-using-rsi: Downloading-data-from-STAC-APIs-using-rsi.html
   articles/How-can-I-: How-can-I-.html
   rsi: rsi.html
-last_built: 2024-07-08T19:34Z
+last_built: 2024-08-15T01:52Z
 urls:
   reference: https://permian-global-research.github.io/rsi/reference
   article: https://permian-global-research.github.io/rsi/articles
