### 2024-08-15 #636 @jhollist
@mdsumner Thank you for your review and worries on the timeline.  We are very much appreciative (and understanding) of the time that all of our reviewers volunteer for rOpenSci!  Keep an eye out for revisions and once those are in, take a look at use our review template (https://devguide.ropensci.org/approval2template.html) to indicate your approval of the revisions or if other changes are needed.

@OldLipe Thank you for your review as well!  As you can see @mikemahoney218 has addressed your reviews.  When you have a chance could you take a look at those and let us know if they address your concerns or if you would like to see some additional changes.  As mentioned above, please use the review template (https://devguide.ropensci.org/approval2template.html) for this.

For both of you, can you provide me with a rough estimate of hours spent on the review?  This is something we keep track of.

Thank you all!
