### 2024-08-16 55ad0d9
Discuss colocated compute
