+++ vignettes/articles/Downloading-data-from-STAC-APIs-using-rsi.Rmd
@@ -355,4 +355,6 @@ If we know how to express our desired query in CQL2, we can write arbitrarily co
 
 rsi is not nearly the only package aiming to help R users take advantage of STAC APIs and build cloud-native geospatial workflows. As shown above, rsi is fundamentally built on top of the excellent rstac package, which I think is a fantastic tool for interactively exploring STAC APIs as well as building and executing queries. My hope is that rsi can provide a useful layer of abstraction over rstac for efficiently downloading assets and performing some of the most common rescaling, masking, and compositing tasks involved in standard data processing workflows.
 
-There are also several other packages which also implement workflows for efficiently accessing and processing data from STAC endpoints, among them the [gdalcubes](https://gdalcubes.github.io/) and [sits](https://e-sensing.github.io/sitsbook/) packages. A core difference between rsi and these packages is that rsi does not have a data model: rsi is focused entirely on finding the bits of data you want from remote endpoints, and getting those bits on your local machine for you to process with your normal spatial data tooling. There are no new classes in rsi (other than the band mapping objects), and the outputs of functions are local rasters. This is an approach that fits better in my head than the more abstract delayed computations in some other packages; at the same time, it's possible that this approach can be less efficient, downloading more data at finer resolutions than is actually needed for a given task. As a result, users need to make sure they're only requesting data they actually need.
+There are also several other packages which also implement workflows for efficiently accessing and processing data from STAC endpoints, among them the [gdalcubes](https://gdalcubes.github.io/) and [sits](https://e-sensing.github.io/sitsbook/) packages. A core difference between rsi and these packages is that rsi does not have a data model: rsi is focused entirely on finding the bits of data you want from remote endpoints, and getting those bits on your local machine for you to process with your normal spatial data tooling. There are no new classes in rsi (other than the band mapping objects), and the outputs of functions are local rasters. This is an approach that fits better in my head than the more abstract delayed computations in some other packages; at the same time, it's possible that this approach can be less efficient, downloading more data at finer resolutions than is actually needed for a given task. 
+
+To make data-downloading functions run faster, make sure you're only requesting the data you actually need -- often restricting the number of assets to be downloaded from each item, specifying a more precise bounding box, or using a narrower date range can cut the amount of data to be downloaded significantly. If there's no way around downloading a large amount of data, consider if running your code on a virtual (cloud) server that's physically closer to your data makes sense for your use case. Many data providers publish information on where their collections are hosted (with cloud providers like Microsoft naming the specific region it's in), and using servers that are closer to the data can result in notably faster download speeds.
