### 2024-08-16 280d513
Assign downloads to objects
