### 2024-08-20 d0acb46
Better check inputs to query function
