### 2024-08-20 7c2c0cf
Add examples of stretching images
