+++ man/get_stac_data.Rd
@@ -396,6 +396,11 @@ landsat_image <- get_landsat_imagery(
   output_filename = tempfile(fileext = ".tif")
 )
 
+landsat_image |> 
+  terra::rast() |>
+  terra::stretch() |>
+  terra::plotRGB()
+
 # The `get_*_imagery()` functions will download 
 # all available "data" assets by default
 # (usually including measured values, and excluding derived bands)
