+++ README.Rmd
@@ -89,7 +89,7 @@ landsat_image <- get_stac_data(
 terra::plot(terra::rast(landsat_image))
 ```
 
-For these common use cases, however, rsi also provides helper functions which
+For common data sets, rsi also provides helper functions which
 provide most of these arguments for you. For instance, that `get_stac_data()`
 call could be as simple as:
 
@@ -101,6 +101,15 @@ landsat_image <- get_landsat_imagery(
   output_filename = tempfile(fileext = ".tif")
 )
 terra::plot(terra::rast(landsat_image))
+```
+
+Note that we've been plotting each band individually so far by calling `terra::plot()`. We could also use `terra::plotRGB()` (after `terra::stretch()`ing the band values) to see what this mosaic of images would look like to the human eye:
+
+```{r}
+landsat_image |> 
+  terra::rast(lyrs = c("R", "G", "B")) |> 
+  terra::stretch() |>
+  terra::plotRGB()
 ```
 
 By default, these functions download data from Microsoft's Planetary Computer
