### 2024-08-20 f95714f
Highlight using st_make_grid()
