+++ R/get_stac_data.R
@@ -222,6 +222,26 @@
 #'   output_filename = tempfile(fileext = ".tif")
 #' )
 #' names(terra::rast(sentinel2_imagery))
+#' 
+#' # If you're downloading data for a particularly large AOI,
+#' # and can't composite the resulting images or want to make
+#' # sure you can continue an interrupted download,
+#' # consider tiling your AOI and requesting each tile separately:
+#' aoi <- sf::st_make_grid(aoi, n = 2)
+#' tiles <- lapply(
+#'   seq_along(aoi),
+#'   function(i) {
+#'     get_landsat_imagery(
+#'       aoi[i],
+#'       start_date = "2022-06-01",
+#'       end_date = "2022-08-30",
+#'       output_filename = tempfile(fileext = ".tif")
+#'     )
+#'   }
+#' )
+#' # You'll get a list of tiles that you can then composite or 
+#' # work with however you wish:
+#' unlist(tiles)
 #'
 #' @export
 get_stac_data <- function(aoi,
