### 2024-09-11 0f2d5e4
Changes for mdsumner review (#78)
* Add rev roles

* Discuss colocated compute

* Clarify downloading files without processing

* Assign downloads to objects

* Better check inputs to query function

* Add examples of stretching images

* Document merge option

* Highlight using st_make_grid()
