+++ vignettes/articles/Downloading-data-from-STAC-APIs-using-rsi.Rmd
@@ -228,12 +228,14 @@ So, with that out of the way, let's actually download and composite some Landsat
 
 ```{r}
 projected_ashe <- sf::st_transform(ashe, 6542)
-get_landsat_imagery(
+ashe_landsat <- get_landsat_imagery(
   aoi = projected_ashe,
   start_date = "2021-06-01",
   end_date = "2021-06-30",
   output_filename = tempfile(fileext = ".tif")
-) |>
+) 
+
+ashe_landsat |>
   terra::rast() |>
   terra::plot()
 ```
@@ -256,14 +258,15 @@ landsat_band_mapping$planetary_computer_v1["coastal"]
 And last but not least, if we wanted to skip compositing altogether, we could set the `composite_function` argument to `NULL` in order to download each image in our spatiotemporal AOI separately:
 
 ```{r}
-get_landsat_imagery(
+ashe_no_composite <- get_landsat_imagery(
   aoi = projected_ashe,
   start_date = "2021-06-01",
   end_date = "2021-06-30",
   output_filename = tempfile(fileext = ".tif"),
   composite_function = NULL,
   mask_function = NULL # otherwise half of these images are blank
-) |>
+) 
+ashe_no_composite |>
   lapply(terra::rast) |>
   lapply(terra::plot) |>
   invisible()
@@ -338,13 +341,14 @@ custom_query_function <- function(bbox,
   rstac::items_fetch(rstac::post_request(request))
 }
 
-get_landsat_imagery(
+ashe_cql <- get_landsat_imagery(
   aoi = projected_ashe,
   start_date = "2021-06-01",
   end_date = "2021-06-30",
   output_filename = tempfile(fileext = ".tif"),
   query_function = custom_query_function
-) |>
+)
+ashe_cql |>
   terra::rast() |>
   terra::plot()
 ```
@@ -355,4 +359,6 @@ If we know how to express our desired query in CQL2, we can write arbitrarily co
 
 rsi is not nearly the only package aiming to help R users take advantage of STAC APIs and build cloud-native geospatial workflows. As shown above, rsi is fundamentally built on top of the excellent rstac package, which I think is a fantastic tool for interactively exploring STAC APIs as well as building and executing queries. My hope is that rsi can provide a useful layer of abstraction over rstac for efficiently downloading assets and performing some of the most common rescaling, masking, and compositing tasks involved in standard data processing workflows.
 
-There are also several other packages which also implement workflows for efficiently accessing and processing data from STAC endpoints, among them the [gdalcubes](https://gdalcubes.github.io/) and [sits](https://e-sensing.github.io/sitsbook/) packages. A core difference between rsi and these packages is that rsi does not have a data model: rsi is focused entirely on finding the bits of data you want from remote endpoints, and getting those bits on your local machine for you to process with your normal spatial data tooling. There are no new classes in rsi (other than the band mapping objects), and the outputs of functions are local rasters. This is an approach that fits better in my head than the more abstract delayed computations in some other packages; at the same time, it's possible that this approach can be less efficient, downloading more data at finer resolutions than is actually needed for a given task. As a result, users need to make sure they're only requesting data they actually need.
[…]
