+++ DESCRIPTION
@@ -4,6 +4,10 @@ Version: 0.2.1.9000
 Authors@R: c(
     person("Michael", "Mahoney", , "mike.mahoney.218@gmail.com", role = c("aut", "cre"),
            comment = c(ORCID = "0000-0003-2402-304X")),
+        person("Felipe", "Carvalho", role = "rev",
+           comment = "Felipe reviewed the package (v. 0.3.0) for rOpenSci, see <https://github.com/ropensci/software-review/issues/636>"),
+    person("Michael", "Sumner", role = "rev",
+           comment = "Michael reviewed the package (v. 0.3.0) for rOpenSci, see <https://github.com/ropensci/software-review/issues/636>"),
     person("Permian Global", role = c("cph", "fnd"))
   )
 Description: Downloads spatial data from spatiotemporal asset catalogs 

[…]
