### 2024-09-17 #636 @mdsumner
Ok amazing, love these responses and the changes you've made.   I think you'll need to link to this issue in the Details section, because it's a really great section on the concerns and the journey you've been on. (I'm getting more enmeshed in the python side via odc).

There's no showstoppers now from my perspective, I think you've responded to this review brilliantly and I'm stoked with all the updates you've made, and the explanations.  Please consider my take as 'approved'. @jhollist

> I'm curious if you have ideas on an efficient way to do this.

I actually didn't mean automated testing validation, just like a real world example where we can get confirmation of the values we see in a small context.   I will follow up when I can but I don't consider this a blocker or anything.

Also I need to follow up here (which I can't remember exactly now, I may have been thinking about a different part of the help content). When I explore again I will bring this up, outside this review as an issue/discussion piece.

> I'm not sure I entirely follow you here! Would you be able to give an example?

Thanks!!
