### 2024-09-18 #636 @jhollist
Looks like we are really close on this one!

@OldLipe do you feel like to @mikemahoney218 has addressed the issues raised in your review?

@mdsumner Thank you for the follow up and the approval!  How many hours do you think you spent on the review?

You both can use this template for your response: https://devguide.ropensci.org/approval2template.html  or you can just provide that directly as well!
