### 2024-09-24 #636 @jhollist
@mdsumner and @OldLipe Just trying to finalize this.  See https://github.com/ropensci/software-review/issues/636#issuecomment-2358634766
