### 2024-09-24 #636 @mdsumner
7 hours was my tally 🙏
