### 2024-10-01 #636 @ropensci-review-bot
Approved! Thanks @mikemahoney218 for submitting and @mdsumner, @OldLipe for your reviews! :grin:
