### 2024-10-01 #636 @mikemahoney218
Thank you _so_ much @jhollist , @mdsumner , and @OldLipe ! This was a fantastic process (as usual with rOpenSci!).

@jhollist -- I mentioned at the start of the review, but I'm not able to transfer this repo to the rOpenSci namespace. I'm not sure which boxes on the checklist still apply -- and I've got the faintest memory that the package needs to get added to a registry somewhere, but I forget where that is. Does this make sense?
