### 2024-10-03 #636 @maelle
:wave: here! https://github.com/ropensci/roregistry/commit/e5d2c0271fd2e66b7d0b42b098b6ac9995992238 should be it but I'll be checking the package and registry building to be sure. :smile_cat:
