### 2024-10-03 #636 @mpadge
Registry now [updated as expected](https://github.com/ropensci/roregistry/blob/41288b188a68b0edf90c344eae600cab19a44dba/registry.json#L332-L338). All good!
