### 2024-10-03 #636 @maelle
https://docs.ropensci.org/rsi/ :tada:
